Refactor PK examples out of crypto and mv non-sig specific to pk dir

+++ pk/rsa/rsa-nb.c
@@ -60,7 +60,7 @@ static int load_file_to_buffer(const char* filename, byte** fileBuf, int* fileLe
     printf("File %s is %d bytes\n", filename, *fileLen);
 
     /* Allocate buffer for image */
-    *fileBuf = malloc(*fileLen);
+    *fileBuf = malloc((size_t) *fileLen);
     if(!*fileBuf) {
         printf("File buffer malloc failed!\n");
         ret = EXIT_FAILURE;
@@ -68,7 +68,7 @@ static int load_file_to_buffer(const char* filename, byte** fileBuf, int* fileLe
     }
 
     /* Load file into buffer */
-    ret = (int)fread(*fileBuf, 1, *fileLen, file);
+    ret = (int)fread(*fileBuf, 1, (size_t) *fileLen, file);
     if(ret != *fileLen) {
         printf("Error reading file! %d", ret);
         ret = EXIT_FAILURE;
